Initial Commit - 12/03/2017: ============================
Initial attempt for a sudoku solver.
Function works for the test file input: incomplete.txt. Output file solution.txt is the correct solution for the input.
However, function does not solve unsolved.txt.

** Class Description **

Classes:
--------
Board: All 81 cells of a sudoku puzzle. Initialises to the values stored in incomplete.txt.
Methods:
full_list- returns a list of all values on the board
update- updates the cell at a certain row and column, with a value

Unit: 9 cells of a unit on the board. Initialises based on incomplete.txt values.
Methods:
update - updates the row, column of a unit with a value
total - returns the sum of all values in the unit
check - checks if a certain value is in the unit
full_list - returns the set of values in the unit
get_num_blanks - returns the number of empty cells in the unit

Row: 9 cells in a row of the board. Initialises based on incomplete.txt values.
Methods:
update - updates a column of the row with a value
total - returns the sum of values in the row
check - returns True if the input value is in the row

Col: 9 cells in a column of the board. Initialises based on incomplete.txt values.
Methods:
update - updates a row of the column with a value
total - returns the sum of values in the column
check - returns True if the input value is in the column

Many functions are contained in the 2 .py files. Each function has a brief description, and code is commented throughout.