Commit 26/03/2017: ==================
New Functionality:
Added a new solver method. Now, when preemptive sets cannot further reduce the range of cells on the Board, a new Board object will be created. A random value will be entered in a cell of this Board.
Based on whether this can lead to a valid solution or not, either a solution will be found, or we revert to the previous Board, with the random value removed from the range of the cell, as all possibilities for this cell have been investigated.
Note, this occurs in a loop, and multiple new Board objects are possible. When a possibility has been deemed invalid, the associated object is deleted.

Functionality seems to be working. All tests seem to be passing. Further testing to follow.